patient-manager/patientManager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-memcached==1.62
django-memcached==0.1.2
python-decouple==3.8
Faker==20.1.0
gunicorn==21.2.0
django-cors-headers==4.3.1